chore(setup):setup_postgres_and sequelize_ORM
changes controller folder into services
creates index file in routes that will hold all routes

[ready for reviews #182201659]
