+++ Procfile
@@ -1,2 +1,1 @@
     
-"make-model": "npx sequelize-cli model:generate --name",
